+++ src/bp/ui-studio/src/web/components/Layout/index.tsx
@@ -1,38 +1,38 @@
-import { lang } from 'botpress/shared'
-import React from 'react'
+import { lang, utils } from 'botpress/shared'
+import React, { FC, Fragment, useEffect, useRef, useState } from 'react'
 import { HotKeys } from 'react-hotkeys'
 import { connect } from 'react-redux'
 import { Redirect, Route, Switch } from 'react-router-dom'
 import SplitPane from 'react-split-pane'
 import { bindActionCreators } from 'redux'
-import { toggleBottomPanel, updateDocumentationModal, viewModeChanged } from '~/actions'
+import { toggleBottomPanel, viewModeChanged } from '~/actions'
 import SelectContentManager from '~/components/Content/Select/Manager'
-import DocumentationModal from '~/components/Layout/DocumentationModal'
 import PluginInjectionSite from '~/components/PluginInjectionSite'
 import BackendToast from '~/components/Util/BackendToast'
-import { isInputFocused } from '~/keyboardShortcuts'
 import Config from '~/views/Config'
 import Content from '~/views/Content'
 import FlowBuilder from '~/views/FlowBuilder'
 import Logs from '~/views/Logs'
 import Module from '~/views/Module'
-import Notifications from '~/views/Notifications'
 import OneFlow from '~/views/OneFlow'
 
 import BotUmountedWarning from './BotUnmountedWarning'
+import CommandPalette from './CommandPalette'
 import GuidedTour from './GuidedTour'
 import LanguageServerHealth from './LangServerHealthWarning'
-import layout from './Layout.styl'
+import layout from './Layout.scss'
 import Sidebar from './Sidebar'
 import StatusBar from './StatusBar'
-import BottomPanel from './StatusBar/BottomPanel'
+import Toolbar from './Toolbar'
+import BottomPanel from './Toolbar/BottomPanel'
+
+const { isInputFocused } = utils
 
 interface ILayoutProps {
   viewModeChanged: any
   viewMode: number
   docModal: any
   docHints: any
-  updateDocumentationModal: any
   location: any
   toggleBottomPanel: () => null
   history: any
@@ -40,193 +40,176 @@ interface ILayoutProps {
   translations: any
 }
 
-class Layout extends React.Component<ILayoutProps> {
-  private botName: string
-  private botId: string
-  private mainEl: HTMLElement
-  private statusBarEmitter: any
+const Layout: FC<ILayoutProps> = props => {
+  const mainElRef = useRef(null)
+  const [langSwitcherOpen, setLangSwitcherOpen] = useState(false)
+  const [guidedTourOpen, setGuidedTourOpen] = useState(false)
 
-  state = {
-    emulatorOpen: false,
-    langSwitcherOpen: false,
-    guidedTourOpen: false
-  }
-
-  componentDidMount() {
-    this.botName = window.BOT_NAME
-    this.botId = window.BOT_ID
-
-    const viewMode = this.props.location.query && this.props.location.query.viewMode
+  useEffect(() => {
+    const viewMode = props.location.query && props.location.query.viewMode
 
     setImmediate(() => {
-      this.props.viewModeChanged(Number(viewMode) || 0)
+      props.viewModeChanged(Number(viewMode) || 0)
     })
 
     setTimeout(() => BotUmountedWarning(), 500)
-  }
+  }, [])
 
-  componentDidUpdate(prevProps) {
-    if (!prevProps.translations && this.props.translations) {
-      lang.extend(this.props.translations)
+  useEffect(() => {
+    if (props.translations) {
+      lang.extend(props.translations)
       lang.init()
     }
-  }
+  }, [props.translations])
 
-  toggleEmulator = () => {
+  const toggleEmulator = () => {
     window.botpressWebChat.sendEvent({ type: 'toggle' })
   }
 
-  toggleGuidedTour = () => {
-    this.setState({ guidedTourOpen: !this.state.guidedTourOpen })
+  const toggleGuidedTour = () => {
+    setGuidedTourOpen(!guidedTourOpen)
   }
 
-  focusEmulator = e => {
+  const focusEmulator = e => {
     if (!isInputFocused() || e.ctrlKey) {
       e.preventDefault()
       window.botpressWebChat.sendEvent({ type: 'show' })
     }
   }
 
-  closeEmulator = e => {
+  const closeEmulator = e => {
     window.botpressWebChat.sendEvent({ type: 'hide' })
   }
 
-  toggleDocs = e => {
+  const toggleDocs = e => {
     e.preventDefault()
-    if (this.props.docModal) {
-      this.props.updateDocumentationModal(null)
-    } else if (this.props.docHints.length) {
-      this.props.updateDocumentationModal(this.props.docHints[0])
+
+    if (props.docHints.length) {
+      window.open(`https://botpress.com/docs/${props.docHints[0]}`, '_blank')
     }
   }
 
-  toggleLangSwitcher = e => {
+  const toggleLangSwitcher = e => {
     e && e.preventDefault()
     if (!isInputFocused()) {
-      const langSwitcherOpen = !this.state.langSwitcherOpen
-      this.setState({ langSwitcherOpen }, () => {
+      setLangSwitcherOpen(() => {
         // lang switcher just closed
         if (!langSwitcherOpen) {
-          this.focusMain()
+          focusMain()
         }
+        return !langSwitcherOpen
       })
     }
   }
 
-  focusMain = () => {
-    if (this.mainEl) {
-      this.mainEl.focus()
+  const focusMain = () => {
+    if (mainElRef?.current) {
+      mainElRef.current.focus()
     }
   }
 
-  gotoUrl = url => {
+  const gotoUrl = url => {
     if (!isInputFocused()) {
-      this.props.history.push(url)
-      this.focusMain()
+      props.history.push(url)
+      focusMain()
     }
   }
 
-  toggleBottomPanel = e => {
+  const toggleBottomPanel = e => {
     e.preventDefault()
-    this.props.toggleBottomPanel()
+    props.toggleBottomPanel()
   }
 
-  goHome = () => {
+  const goHome = () => {
     if (!isInputFocused()) {
       window.location.href = `${window.ROOT_PATH}/admin`
     }
   }
 
-  render() {
-    if (this.props.viewMode < 0 || !this.props.translations) {
-      return null
-    }
+  if (props.viewMode < 0 || !props.translations) {
+    return null
+  }
 
-    const keyHandlers = {
-      'emulator-focus': this.focusEmulator,
-      cancel: this.closeEmulator,
-      'docs-toggle': this.toggleDocs,
-      'bottom-bar': this.toggleBottomPanel,
-      'lang-switcher': this.toggleLangSwitcher,
-      'go-flow': () => this.gotoUrl('/flows'),
-      'go-home': this.goHome,
-      'go-content': () => this.gotoUrl('/content'),
-      'go-module-code': () => this.gotoUrl('/modules/code-editor'),
-      'go-module-qna': () => this.gotoUrl('/modules/qna'),
-      'go-module-testing': () => this.gotoUrl('/modules/testing'),
-      'go-module-analytics': () => this.gotoUrl('/modules/analytics'),
-      'go-understanding': () => this.gotoUrl('/modules/nlu')
-    }
+  const keyHandlers = {
+    'emulator-focus': focusEmulator,
+    cancel: closeEmulator,
+    'docs-toggle': toggleDocs,
+    'bottom-bar': toggleBottomPanel,
+    'lang-switcher': toggleLangSwitcher,
+    'go-flow': () => gotoUrl('/flows'),
+    'go-home': goHome,
+    'go-content': () => gotoUrl('/content'),
+    'go-module-code': () => gotoUrl('/modules/code-editor'),
+    'go-module-qna': () => gotoUrl('/modules/qna'),
+    'go-module-testing': () => gotoUrl('/modules/testing'),
+    'go-module-analytics': () => gotoUrl('/modules/analytics'),
+    'go-understanding': () => gotoUrl('/modules/nlu')
+  }
 
-    const splitPanelLastSizeKey = `bp::${window.BOT_ID}::bottom-panel-size`
-    const lastSize = parseInt(localStorage.getItem(splitPanelLastSizeKey) || '175', 10)
-    const bottomBarSize = this.props.bottomPanel ? lastSize : '100%'
+  const splitPanelLastSizeKey = `bp::${window.BOT_ID}::bottom-panel-size`
+  const lastSize = parseInt(localStorage.getItem(splitPanelLastSizeKey) || '175', 10)
+  const bottomBarSize = props.bottomPanel ? lastSize : '100%'
 
-    return (
-      <HotKeys handlers={keyHandlers} id="mainLayout">
-        <DocumentationModal />
-        <div style={{ display: 'flex' }} className={layout.container}>
-          <Sidebar />
+  return (
+    <Fragment>
+      <HotKeys handlers={keyHandlers} id="mainLayout" className={layout.mainLayout}>
+        <Sidebar />
+        <div className={layout.container}>
+          <Toolbar
+            hasDoc={props.docHints?.length}
+            toggleDocs={toggleDocs}
+            onToggleEmulator={toggleEmulator}
+            toggleBottomPanel={props.toggleBottomPanel}
+          />
           <SplitPane
             split={'horizontal'}
             defaultSize={lastSize}
             onChange={size => size > 100 && localStorage.setItem(splitPanelLastSizeKey, size.toString())}
             size={bottomBarSize}
             maxSize={-100}
-            pane1Style={
-              this.props.bottomPanel
-                ? {
-                    maxHeight: 'calc(100% - 100px)'
-                  }
-                : null
-            }
+            className={layout.mainSplitPaneWToolbar}
           >
-            <div>
-              <main ref={el => (this.mainEl = el)} className={layout.main} id="main" tabIndex={9999}>
-                <Switch>
-                  <Route
-                    exact
-                    path="/"
-                    render={() => {
-                      if (!window.IS_BOT_MOUNTED) {
-                        return <Redirect to="/config" />
-                      }
+            <main ref={mainElRef} className={layout.main} id="main" tabIndex={9999}>
+              <Switch>
+                <Route
+                  exact
+                  path="/"
+                  render={() => {
+                    if (!window.IS_BOT_MOUNTED) {
+                      return <Redirect to="/config" />
+                    }
 
-                      return window.USE_ONEFLOW ? <Redirect to="/oneflow" /> : <Redirect to="/flows" />
-                    }}
-                  />
-                  <Route exact path="/content" component={Content} />
-                  <Route exact path="/flows/:flow*" component={FlowBuilder} />
-                  <Route exact path="/config" component={Config} />
-                  <Route exact path="/oneflow/:flow*" component={OneFlow} />
-                  <Route exact path="/modules/:moduleName/:componentName?" render={props => <Module {...props} />} />
-                  <Route exact path="/notifications" component={Notifications} />
-                  <Route exact path="/logs" component={Logs} />
-                </Switch>
-              </main>
-            </div>
+                    return window.USE_ONEFLOW ? <Redirect to="/oneflow" /> : <Redirect to="/flows" />
+                  }}
+                />
+                <Route exact path="/content" component={Content} />
+                <Route exact path="/flows/:flow*" component={FlowBuilder} />
+                <Route exact path="/config" component={Config} />
+                <Route exact path="/oneflow/:flow*" component={OneFlow} />
+                <Route exact path="/modules/:moduleName/:componentName?" render={props => <Module {...props} />} />
+                <Route exact path="/logs" component={Logs} />
+              </Switch>
+            </main>
             <BottomPanel />
           </SplitPane>
 
           <PluginInjectionSite site="overlay" />
           <BackendToast />
           <SelectContentManager />
-
-          <StatusBar
-            botName={this.botName || this.botId}
-            onToggleEmulator={this.toggleEmulator}
-            emitter={this.statusBarEmitter}
-            langSwitcherOpen={this.state.langSwitcherOpen}
-            toggleLangSwitcher={this.toggleLangSwitcher}
-            onToggleGuidedTour={this.toggleGuidedTour}
-            toggleBottomPanel={this.props.toggleBottomPanel}
-          />
-          <GuidedTour isDisplayed={this.state.guidedTourOpen} onToggle={this.toggleGuidedTour} />
+          <GuidedTour isDisplayed={guidedTourOpen} onToggle={toggleGuidedTour} />
           <LanguageServerHealth />
         </div>
       </HotKeys>
-    )
-  }
+      <StatusBar
+        onToggleEmulator={toggleEmulator}
+        langSwitcherOpen={langSwitcherOpen}
+        toggleLangSwitcher={toggleLangSwitcher}
+        onToggleGuidedTour={toggleGuidedTour}
+        toggleBottomPanel={props.toggleBottomPanel}
+      />
+      <CommandPalette toggleEmulator={toggleEmulator} />
+    </Fragment>
+  )
 }
 
 const mapStateToProps = state => ({
@@ -236,7 +219,6 @@ const mapStateToProps = state => ({
   translations: state.language.translations
 })
 
-const mapDispatchToProps = dispatch =>
-  bindActionCreators({ viewModeChanged, updateDocumentationModal, toggleBottomPanel }, dispatch)
+const mapDispatchToProps = dispatch => bindActionCreators({ viewModeChanged, toggleBottomPanel }, dispatch)
 
 export default connect(mapStateToProps, mapDispatchToProps)(Layout)